requirements.txt
django
django-environ
django-filter
django-probes
django-rest-framework
psycopg2-binary
uwsgi

# test
black
flake8
isort
mypy
pytest
pytest-cov
pytest-django
pytest-freezegun

# dev
ipdb
pip-check
